output() and friends now support scalar references.
This will make life a lot easier while testing itself.  No more having to
screw around with TieOut.  Had to slurp in IO::Scalar for 5.6 support.
I started with just TieOut but getting the self-tie right was too much
of a PITA so I figured just go with something that works.