Make Result accessors return the object on set.
This allows this syntax:

    my $result = $Builder->ok($test, $name)
                   ->diag({...});

Unfortunately, Mouse is hard wired to return $self->{$key} from its
accessors so I had to throw that all out.  This also means I lost the
ability to easily set defaults for accessors.  Now more is hard
wired than I'd like.  On the whole Result is a mess.

I'm considering junking the whole subclass factory for the moment.