Change results so different types are different classes.
This is a huge refactoring, pairing with rafl.

Each type of result is now a class composed of various result roles like
"pass" and "todo".

TB2::Result is now a result factory.  Instead of passing in the type of the
result directly, you hand it the result attributes and it figures it out.

Result objects were going to be immutable, but I decided it was too unperlish
and they need to be changed anyway.  The essential aspects of a result,
if it passed, if it was a skip, is inherently immutable.