Add the rest of the "event" methods.
plan() becomes stream_start() and set_plan()
done_testing() becomes stream_end()
history() is split into a history() fetcher and a set_history() setter to
allow a hook when the history is changed.
Same for the formatter.

This allows the Test::NoWarnings sample to work properly no matter when the
number of tests is set (at the beginning or end of the test or not at all)