Implement the use case of writing is().
Necessitated changing when the result is formatted to happen in
the top-most assert_end() to allow wrapper functions, such as is(),
an opportunity to add to the result before formatting.

Added TB2->at_top and TB2->in_assert helper functions to let
asserts better know where they are in the stack.

TB2->ok() is now a proper test function which calls assert_start
and assert_end.

Test functions must return a single result object because we need
to pass it into assert_end for formatting.  If you really want
your user-visible test function to return something else, wrap it.