Put the counter into the Formatter
The counter was removed from History.  This puts it back into the TAP
formatter.  Other formatters probably won't have a counter.

It also fixes tests wrt the removal of History->should_keep_history.
That's covered by NoHistory.

It also fixes tests wrt the removal of History->summary.  Its not
really needed now that History has stats.

The Formatter will now EITHER use the test_number on a Result OR
it will increment and use the counter.  Not both.  This lets a Builder
tell the formatter to use a specific number once without changing the
count.

The tests still don't 100% pass.  Working on it.