Pass along the EventCoordinator to accept_event/result when posting.
This will allow EventWatchers to issue their own events which will become necessary shortly.

The obvious way of doing this, by storing the EventCoordinator in the EventWatcher creates
a circular reference and thus will leak memory.  It also only allows a watcher to be
attached to one coordinator at a time... which we don't need right now but might be
handy.