Have Test::Builder start the stream inside the individual plan methods
Things expect to be able to call expected_tests() and no_plan() and skip_all() alone
without having called plan.  This requires them to start the stream.

It also keeps Test::Builder from outputting the TAP version header even if it's going to
die with a bad plan.

This fixes Test::Class.

Fixes #97