Fix TB1 so it works without a formatter.  This lets TB2::Tester work.
There are two hacks here.  First, the Formatter is replaced with a BlackHole object.
This avoids TB1 (and other builders) from having to check that the formatter is there so
long as its functionality is optional.

Second, changing the current test with TB1 relies on having a counter.  So we fake one
using the history object.  TB1 will not remember the count has changed.  But that's ok,
only the TAP formatter cares about that and there is no TAP formatter.

Closes #81