Make child() use create().
child() was improperly creating it's own object and calling reset() manually and then copying
filehandles from the parent directly rather than going through the accessors.  Now it uses
create() and the accessors.

Related... fix Test::Builder::NoOutput->create so it doesn't use global filehandles and blow
over them every time create() is called (as by a child).

The reset_outputs() in the basic subtest test were a symptom of this confusion and are
no longer needed.

For #145