Add basic subtest handling to the TAP formatter.
It indents the subtest and generates the extra result.

Generating the result is incomplete, the logic in History to determine if the History represents
passing test will be improved.

The subtest formatter should always inherit the streamer, that should be made default for
all formatters.

SubtestStart will need to have a name.

For #170