Adapt Test::Builder->done_testing to work with OnlyOnePlan.
Eliminate $builder->{Done_Testing} internal state.

The double plan error messages have improved and become more generic.

Doing the same test twice is ok... but that may or may not last.

Calling done_testing() with no tests having been run is now an error, before
it was a "1..0" and a 255 exit code.  This may or may not last.

For #184