Get another basic subtest test working.
Strip out most of the junk in subtest().  It was only necessary to ensure the TB object is
restored in case the subtest dies.  That's no longer necessary.

Start the test if it isn't already so the TAP version header comes out before the first
subtest.

Pass context into the subtest events.

Unfortunately, Perl can't seem to see the line where ok() is called from in caller().
It only sees the end of the subtest subroutine.  Not much to be done about that.

For #178