Fix the line numbering in results in subtests.
Turns out I was going way too *high* and the tests in t/subtest/basic.t had the wrong
level set.  Test::Builder defaults to a level of 1, if you use it directly it has to be
set to 0.

There's still busted tests in t/subtest/line_numbers.t

For #178