Remove the bits of Test::Builder's own filehandles.
* A separate todo output handle no longer exists as a concept.  It was never very
  useful and just an extra thing for people to forget to set when changing outputs.
  It is now just a pass through for $builder->output.  Setting it does nothing.

Fix Test::Builder::Tester to not worry about the todo_output.

Plug Test::Builder->reset_outputs into TB2::Formatter->reset_streamer.

The reset_outputs test was changing the outputs as it was storing their originals.
This messed up because todo_output() is now an alias for output().  It would
record and change output(), then go to record todo_output() but it was already
changed by output().  Do the recording and changing separately to avoid confusion.

For #231