Emulate an accidental behavior of Test::Builder::Tester.
Previously, TBT would not completely reset the test state so the
plan would still be set.  Some tests accidentally relied on this,
anything that would set the plan if it hasn't already been set like
Test::HasVersion or Test::DistManifest.

This change sets a plan in the test test if there was one before
to emulate the behavior.

For #327 and #328