Further emulation of undocumented TBT features.
Tests rely on these TBT 0.9x quirks...
* Copy the existing plan
* Set no_ending

Also, move setting these things *after* setting up the new state to avoid
them leaking out.  This works for no_ending, but no_change_exit_code is
still part of the builder.  But if that gets put into a handler this code
will be ready for it.

Test::Builder::Tester is now discouraged.  Testing the formatted TAP is just
asking for incompatibilities.  This frees us to make it do all the same
quirks as 0.9x and not worry about trying to improve it (and breaking things).
It can now just be a transitional thing.