The basic fork test runs, but they're totally out of sync.
Added TB2::SyncStore and the coordiante_forks option.

While the sync directories are the same, it looks like each process is
using its own file.  Possibly an overly lazy default.  Probably best
if the mapping of an object ID to a filename was deterministic, possibly
using a hash algorithm.

Checking this in because there's a lot of implementation here and
it's getting scary keeping this in my stash.