Issue the test_start event automatically.
Most uses were of the form $self->test_start if !$history->in_test which made it
pointless.  However, a test_start event is still useful as pointed out by both
the TAP and POSIX formatters.  So the event is issued automatically by the
event coordinator.

This also solves a problem with forking.  If you fork before any events are seen,
the builder in each process will try to start the test.  Because fork coordination
is done inside TestState->post_event the builders don't know to check if somebody
else has started the test.  TestState is the only thing which knows about that.
By making starting the test the EventCoordiantor's problem, it neatly avoids
Builders having to know about fork coordination.