Allow event storage to be turned on and off in the middle of a test.
This makes it much easier for test authors to turn event storage on
and off, otherwise it has to be done before anyone issues an event.

Updated the docs to make it clearer this is done at their own risk,
that events may be lost, and to encourge the statistical methods.