like() and unlike() no longer warn about undef.
Not much sense, they can see it in the diagnostics, and the warning
message is a bit confusing.

    Use of uninitialized value $this in pattern match (m//) at x:\test.t line 6.

It reports the right location, but there's no $this from the user's
perspective.

This came in because like used to supress all warnings.  That was
fixed in 8b2f0c6e1fb789b0ea5ed217b5be18705bbe6b8c but it let
uninit warnings slip in.

For #335.