Allow Test::Builder::Tester to use regex matching
TBT has had an undocumented version of this ability for ten years.
This commit improves the interface slightly, documents it, and
fixes the color highlighting to work.

This is important because even with plain old strings we ran into
broken color highlighting! This was a result of magically changing
standard failing messages into regexes internally to cope with the
change in output format.

Of course, we wouldn't need *any* of this if we used a more
modern library (i.e. one that wasn't deprecated) and fragile to
change, but there you go.