Skip thread tests on 5.10.0, they reproducably segfault
fixes #426

Initially I thought this was a PerlIO+Threads interaction in 5.10.0, the
5.10.1 delta mentions such a bug. However the thread segfaults still
happen when I remove PerlIO completely.

In fact I can say for sure this has nothing at all to do with
Test::Builder. I put the following code into a script and ran it on a
fresh 5.10.0, it does not load Test::Builder at all, but still
segfaults:

        use threads;
        threads->create(sub { print "XXX\n" })->join;

It does not get more definitive than that.