Do not make utf8 default afterall, too many platform specific issues
utf8 is no longer the default for the alphas. It might be great for
linux, but some platfroms (windows) just can't do it right. (or I am too
stupid to make it work).

This makes it so that the default is to just use what is already there
for output. It also makes it so that specifying 'utf8' or 'encoding'
when you use Test::More implies 'modern', this avoids the need to
specify both 'modern' and 'utf8'.

This should be sufficient, we cannot use it as a default assumption, but
ti is still encredibly easy to use it.