Removed the indent for BAIL_OUT
In the latest stable version, when we call the BAIL_OUT in the subtests, indentation does not.

I think if we call the BAIL_OUT, test because it is immediately stopped, it is better to be displayed
in no indentation is easy to understand is BAIL_OUT regardless of the state of the sub-test.