Fix ribasushi's thread test
I marked this as skip_all broken a month or so ago, and asked ribasushi
to clarify it. He said he would and then it seemed to get lost in his
todo list. He recently opened ticket 461 to rekindle the issue.

This should fix the issue. This preserves backwards compatability with
at least one workaround used to bypass broken subtest in threaded code.
The issue this works around appears to actually be fixed, but we must
not break things that have to work on old versions.

Fixes #461