Add tests for #388
This apprears to be fixed. I tested in 5.8.9, 5.16.2, and 5.20.1. It
also appears have been fixed in stable at some point as I was not able
to reproduce there either on any perl version.

fixes #388

Marking fixed for now, reopen if you can provide a reliable way to
reproduce this issue (assuming it still occurs for you)