Initial work on an API module
This module WRAPS all the common things people want to do into a single
module. Many of these things are spread around other modules and may not
be easy to find. Here is a single place to go to find these things!

This module also adds sugar to encourage its use.

TODO: Documentation and tests.