Make Context+Event stuff less magic
Fixed #555

This removes the magic event-methods from Context and instead adds
send_event(...). There are still shortcut methods for the main events,
but they are not magic. This also removes the autoload.

(Note: Finished this yesterday, but forgot to commit+push, doing so now
so someone can look at it)