Better behavior for forking inside a subtest
In legacy Test::Builder there was no forking support. When I added it I
defined a behavior for forking in subtests, that behavior was to fork
the subtest. This behavior is not useful, and even the unit test for it
said so. Since writing the behavior into Test::Stream I have come to
realize that the new behavior I defined was not very useful.

This change makes it so that forking in a subtest has a behavior that
should be unsuprising to anyone. When you fork inside a subtest, events
from both processes will go to the subtest. When your subtest ends it
will automatically grab results from any child processes. If results are
generated in a child process after the subtest ends, it generates a lot
of noise to tell you something is wrong.