Simple send unless complex is needed
This adds a simple send method that covers the majority of uses of
Test::Builder.

When subtests, forking, threading, or event hooks are used the method is
swapped for one that takes care of them.

The goal here is to make it MUCH faster when you do not use these extra
features.