IPC improvements and fixes
 * Make it so that the END block frees the remaining hubs so that they
   can clean up.
 * Add helpful diag when files are leftover in IPC
 * Fix race condition
 * Fix load ordering

If you fork/thread before running any tests the root hub would not be
initialized, this resulted in each process creating a root hub.

This change makes it so that if both context and IPC are loaded the root
hub gets initialized. This is done in such a way as to make it work
regardless of load order.