Add the DeepCheck tools
These tools are for checking deep data structures. These will also be
used under the hood by the test-testing tool in the next couple commits.

You will notice a lack of tests in the commit, that is because the tests
require other tools that are to be added as seperate commits. All the
tests will be added as a final commit.