Test-Builder TAP compatability
In an early stage of refactoring, there weas a chaneg to Test::Builder
where a 'print ... "\n"' was changed to a $self->diag("\n"), this
resulted in test failures that were fixed by changing how newlines in
diag/note were handled.

The changes to how the renderign was handled caused all tests to pass,
and allowed all known downstream modules on cpan to pass. Nothing seemed
to care much about the trailing whitespace, or if newlines had a '#'
prefix.

During p5p output testing this was discovered as it resulted in a direct
output comparison.

This commit reverts the mistake and its consequences. DIAG and NOTE
rendering match eahcother better, and are simpler. Several test cases
could be removed as a result (they tested for bad behavior).

A new test has been added that checks that Note, Diag, and the special
legacy-compatible Ok all behave as they did in Test::Builder. This test
should help avoid future breakages.