Put state back into the hub.
Originally the State object and the Hub object were one object. The
state was moved into its own object for an old feature that turned out
to be bad adea. The feature was removed, but the state object was never
squashed back into the hub.

This commit puts the state back into the hub. There is no reason
currently for them to be seperate. The state object is never used
outside of the hub, and nothing ever replaces the hubs state object.

This also allows for the removal of some complicated logic between the
hub and the state that can be made simpler when they are combined. Such
as finish/finalize, and Hub->process calling
$event->update_state($state).

Now instead of having several event objects with custom update_state()
methods, there are 3 functions that events can override to achieve the
same behavior:

 * increments_count() - For tihngs like Ok that bump the test count
 * causes_fail()      - For events that cause a failure
 * callback()         - For the rare event that needs more
                        (plan and bailout)

Overall this results in a significant drop in method calls, and prevents
logic fragmentation.