Fully seperate Ok and Diag
Initially Ok was given diagnostics messages so that the die_on_fail
functionality could be implemented. In fact context->release has long
since solved the problem in a better way. Having the diagnostics inside
the Ok is now an artifact, and one that is causing problems.

Having diagnostics in Ok means duplicating logic such as "which handle
do I use, is this todo?". It also faisl to maintain a seperation of
concerns.