Make third-party meta-data interface consistent
For events, hubs, and contexts, it can be very useful to attach third
party meta-data to the instances. Instead of people simply adding their
own hash keys, or doing it inside-out, or even subclassing all types, we
provide a consistent, and safe interface for attaching the data.

2 of these 3 classes already had this functionality, but it was not
consistent beetween them. Now that it needs to be added to Context as
well I decided to make it consistent in all 3. It is also now
encapsulated by its own package for better maintainability and testing.