Filename portability for Test2.
Explicitly concatenating paths with a forward slash doesn't work
so well on filesystems that don't use slashes as separators. So
use catfile instead.

Also, on VMS, all files have an extension and a zero-length
extension is returned as a '.' from readdir(), so only match
the start of the extension-less hub files when testing for them.