update t/Legacy/Regression/637.t
This test is problematic on some platforms. For now I am making this an
AUTHOR_TESTING test.

I have also changed it to load Test2::IPC which should give it some
extra safety, but I really do not feel that this test needs to run
everywhere, so long as I run it locally before release.