Use pre_filter instead of filter in Test::Builders TODO logic
Fixes #683

Test::Builder uses a hook to attach the TODO status to an event as it
passes through the hub. The 'filter' hook runs in the destination hub,
as a result it was not able to find the TODO status for events generated
in a child process or thread.

This changes the hook to use pre_filter, which runs in the child
process/thread before the event reaches the parent, this allows the TODO
status to be found.