IPC: "Fix" win32 file operations
Win32 can claim it finished writing/renaming/unlinking before actually
finishing, in which case it holds a lock. When there is a sequence of
such events the lock can trip us up, this will make it try a couple
times on win32.

I hope we can find a better way to do this in the future.