Bag: Clean up diagnostics
I think the diagnostics for bag were too noisy. This will simply list
which bag items were not found. If end() is also used it will show items
that were not in the bag, which is another way to give you all the info,
but in a less noisy way.