Change Test2::Plugin::UTF8 to not alter STDOUT/ERR.
Test libraries should avoid altering the global environment lest
code work in testing and not in production.

I also changed the docs to be less specific about what exactly this
module does to allow us flexibility to make small changes in the
future. If you say "this is the same as" someone is going to take
you literally.

See discussion in #92.