handle the minus on negative pids in forks on windows
On windows, forked processes are threads, and reading from $$ gives the
thread id instead, which is a negative number. This interferes with the
field separator in filenames, so this replaces the separator - with one
retrieved from a Util function, thus also enabling later quick changing if
portability issues crop up.