when a subtest forks, do not finish up in both processes
Given this sample program:

  use strict;
  use warnings;
  use Test::More;
  use Test2::API;

  Test2::API::run_subtest("forking subtest" => sub {
    if (fork) {
      wait;
      cmp_ok($?, '!=', 0, "subprocess died");
    } else {
      die "123";
    };
  });

  done_testing;

...we'll see the "forking subtest" fail in both processes.  Also, the
subprocess will now exit 0, so the cmp_ok fails.  This is not the same
behavior as Test::Builder's subtests.  Test::Builder tracks whether the
process has forked and, if it has, dies in the child and only finalizes
the subtest in the parent.  This patch brings that behavior to
Test2::API::run_subtest.

Note that Test::Builder also checks whether Test::Sync::IPC has been
loaded before dying in the subtest.  I have not carried that condition
forward, because I can't find that module anywhere.