Improve the fix to #637
Yeah, #637 was a long time ago...

I recently starting working on a branch where a couple things changed.
These changes revealed a race condition in 637.t where the test would
often pass, but occasionally it would fail saying that a hub can only be
dropped by the pid/thread it was created in.

The original fix to 637 would change the pid/tid inside the hub in the
child thread... problem is that both threads woudl then fight over which
would drop the hub. The original thread pretty much always won, and the
second thread would not have time to report any issues as the thread
would simply end.

With my changes it was apparently no longer the case that the initial
thread always won, so the test would fail off an on.

The fix to the problem si to have the hub not just update it's pid/tid,
but to also update its hid and add itself to the ipc system. This is the
correct way to handle this. The reason this never cropped up before is
because the use case of 637 is pretty obscure and rare.