Mark gh_16 AUTHOR_TESTING on version lower than 5.20
This test does something crazy. The check is important, but I keep
seeing false priblems with it that are hard to correct. I am making sure
it always runs for 5.20+, and also always when AUTHOR_TESTING is set.

Fact is most people are unlikely to have this problem, and the fix seems
to be pretty solid anyway. The problems here are that the test is kind
of crazy, specially for old perl versions that have other exception/END
bugs.