Fix event out-of-order issue
On some filesystems, under some kernels (mainly seen in windows and
freebsd) events had the potential to be read out of order. This was
really a race condition between one process writing files and another
reading them. The reader always sorted events, but occasionally it might
see a later event without seeing an earlier one.

This commit makes several changes:

1) Event files now have a counter starting at 1, and each pid+tid+hid
combo must start at 1 and be sequential. If an event is missing the
system will wait until it arrives before processing the events that
follow it (with a timeout to prevent infinite hanging).

2) An abort by the IPC system will print "bail out!" which will
harnesses know something very bad happened.

3) Aborts add an 'ABORT' file that will cause any ipc cull to exit
immedietly.

4) t/Test2/regression/ipc_files_abort_exit.t has been cleaned up, but is
now only an AUTHOR_TESTING test, it is not necessary for all consumers
of this module to run that test.

5) A hook has been added to IPC::Drivers allowing them to hook intot he
abort process.