Make it easy/fast to add UUID tracking
This makes it easy to add a UUID generator that will be used to
automatically tag every hub, context, and event (once it is processed by
a hub).

It was added this way (instead of via other hooks) for performance
reasons. The generator gets called for every context, and every event,
that is a lot!