Add 'table' support in INFO + Formatters
This adds support for table data in INFO facets and in the TAP
formatter. This also allows other formatters to adopt table support, but
does not break any existing formatters as there is a fallback rendering
to details. Tools need to be updated to add table data if they desire.

This is needed before
https://github.com/Test-More/Test2-Suite/issues/127 can be fixed.