Fix TODO = '' bug
Previous fix for #812 was incorrect.
Fixes #829

The original issue is that a TODO = '' would result in a failing test,
but diagnostics that report TODO

MY original fix was to make TODO = '' set todo, which was incorrect

The correct fix was to make the diagnostics not incorrectly report a
TODO when it is set to ''.

To add consfusion, Test::Builder always ignores TODO = '', but does NOT
ignore TB->set_todo('').... That is a legacy behavior that must be
preserved :-(