Fallback to Data::Dumper or simple "sorry" if JSON::PP is not installed
This codepath is only hit when something goes horribly wrong. JSON::PP
is loaded to help provide a better error message. Now if JSON::PP is not
installed (5.10 and below) it will fallback to Data::Dumper.

If everything fails to dump the data then it will say it is sorry.

Fixes #851