Fix #862, Test::More's $TODO inside intercept
Test::More's use of a $TODO variable that is localized to a defined
value to set the state is problematic. Test::Builder, when loaded, adds
the todo callback to the hub, and it is always called, even when $TODO
is not set, it just does nothing in that case. When $TODO is set is does
the same thing as the Test2 tools and adds amnesty.

intercept intentionally creates a new
hub that is not based on whatever is at the top of the stack. It is like
this specifically so that it is a virgin state in which you can test
things like hub callbacks. $TODO from Test::More does not work because
the callback has not been added again for the intercepte hub.

This is a non-issue when using intercept to test Test2 tools, they do
not use global variables that get localized. The design specifically
avoids such things. This is only a problem when you are using Test2's
intercept() to test Test::Builder based tools.

The problem is, simply having intercept inherit the Test::Builder todo
callback would break the guarantee that tools tested in intercept() get
a virgin hub with no other tools contamination. I added a new way to
specify something should be inherited across intercept, but it has
callbacks to clean/restore state such as global variables. This allows a
tool to say "Inherit me, I will make sure the encironment is clean".

Test::Builder's todo hook now inherits, and clears/restores $TODO across
the intercept boundry.

Note: The Test2::Tools::Grab in Test2::Suite also needs to be updated.