Improve the API around intercept {}
This updates intercept {} so that it returns a blessed arrayref that can
be used to filter/summarize/validate the events. This makes validating
test results and facets much easier.

Drop 5.6 from travis for now

This change breaks tests on 5.6. I do not personally support 5.6 and
will not invest in fixing this. However if someone wants to submit a
patch that fixes 5.6 and re-enables it in travis I will accept the
patch, so long as it does not cause other issues.