avoid closing over scalar in BEGIN block in cmp_ok eval
When comparing values using cmp_ok, a string eval is used to check using
the correct op. It also carries through the caller's warnings. If the
BEGIN inside the eval closes over a variable, it slightly alters how the
eval works. This leads to Devel::Cover tracking it like a sub, which
means it holds on to both the op tree, and the values it closes over.
That includes the values that are being compared. This means that
any references passed to Test::More::is would not be destroyed until
process end.

Avoid this by converting the warnings bits value to a string, and
inlining that in the generated code.