avoid using ' package separator for isn't function
Test::More provides the isnt() function. Since perl allows single quotes
to be used as package separators, this function is also "cleverly" made
available as isn't(). This is exactly equivalent to installing it as
isn::t.

In the future, perl may deprecate or remove single quote package
separators, so stop using the "isn't" form in our code. The function is
still installed in the same location, so users are free to continue
using it, putting the compatibility burden on them rather than
Test::More.