Support T2_AUTO_DUMP and T2_AUTO_DEPARSE environment variables.
T2_AUTO_DUMP can be set to true to automatically dump $got if
an is/isnt/like/unlike test fails. If it's set to a module name,
and the module offers a `Dump` method, this is used instead of
Data::Dumper.

If T2_AUTO_DEPARSE is set to true, then $Data::Dumper::Deparse
will be set to true.