More.pm - make isn::t() produce deprecation warnings
Use of apostrophe as a package separator was deprecated in Perl 5.37.9,
and will be removed outright by Perl 5.42.0. So rework isn::t() to produce
warnings when this happens. The tests do not actually use the apostrophe
spelling, so this is safe. The warning message can be silence by disabling
deprecated warnings