Test2::API - harden T2_FORMATTER env access.
There were two problems. The first is accessing the %ENV hash when it
might locked. Using Test2::Util::_env_get() solves that problem.

The second issue is that the patterns might match "+" with no extra string,
leaving the formatter as the empty string. The third problem is the
defaulting for the formatter could end up not being triggered and the
formatter being left as the empty string.

This fixes all three issues. I noticed the second two when I fixing the
first and did it wrong accidentally.