Fix #882
This will "fix" or more accurately address #882. Theoretically the
situation should not occur at all. In practice when things are messy, or
some poorly written tool messes with Test2::API::Stack or
Test2::API::Hub then it can happen. In these cases catch it and provide
an intentional error instead of a bunch of unhelpful warnings