is_deeply: fix handling of VSTRING and LVALUE refs
VSTRING and LVALUE are possible return types from ref, but they are
essentially special cases of SCALAR refs. As far as is_deeply is
concerned, they should be treated as equivalent.

Update the _type function to normalize VSTRING and LVALUE into SCALAR,
so that the deep checks properly treat them as equivalent.

This fixes errors from doing comparisons with LVALUE refs, as well as
fixing VSTRINGs comparing equal to an equivalent normal SCALAR string.