Test2::Hub - fix remove_context_{aquire,init,release} logic
As far as I can tell, the intended logic is:

 1. Build a set from passed-in arguments, represented as a hash.
 2. Keep only context handlers that are not elements of the set (i.e.
    not keys of the hash).

However, what the code actually did was to compare the result of `!`
(which is a boolean, numerically 0 or 1) to a coderef, which is always
false. In other words, the remove_context_* methods would ignore their
arguments and always remove everything.