Make it possible to queue up diag for the context
fixes #1004

This adds a system for queuing diag messages that will be attached to
the next context that releases, upon release the diags will be issued so
long as the context caused at least 1 failure. The diags will be
disgarded if the context did not add any failures.

the `no_warnings` and `lives` tools both now automatically add the
warning/exception to the diag queue.

Examples:

This will add the warning as a diagnostics message
ok(no_warnings { warn "xxx" }, "Did not get any warnings");

This will not add the warning-diag as no failures are caused
ok(!no_warnings { warn "xxx" }, "Got warnings");

This will add the exception as a diagnostics message
ok(lives { die "XXX" }, "Did not die");

This will not add the exception as a diagnostics message
ok(!lives { die "XXX" }, "Died");